+++ lib/Service/DataloadService.php
@@ -130,10 +130,10 @@ class DataloadService
     public function execute(int $dataloadId)
     {
         $dataloadMetadata = $this->DataloadMapper->getDataloadById($dataloadId);
+        $option = json_decode($dataloadMetadata['option'], true);
         $result = $this->getDataFromDatasource($dataloadId);
         $insert = $update = 0;
         $datasetId = $result['datasetId'];
-        $option = json_decode($dataloadMetadata['option'], true);
 
         if (isset($option['delete']) and $option['delete'] === 'true') {
             $this->StorageController->delete($datasetId, '*', '*');