+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - SQL error for shared reports [#98](https://github.com/rello/analytics/issues/98)
 - Datasource options field in DB too short [#78](https://github.com/rello/analytics/issues/78)
 - JSON source not adding timestamp [#106](https://github.com/rello/analytics/issues/106)
+- Error in occ- and scheduled load [#109](https://github.com/rello/analytics/issues/109)
 
 ## 3.2.0 - 2021-01-02
 ### Added